Support for large number of edges. Using long long instead of int/long for some variables.

+++ graph.cpp
@@ -95,8 +95,8 @@ template <typename captype, typename tcaptype, typename flowtype>
 template <typename captype, typename tcaptype, typename flowtype> 
 	void Graph<captype,tcaptype,flowtype>::reallocate_arcs()
 {
-	int arc_num_max = (int)(arc_max - arcs);
-	int arc_num = (int)(arc_last - arcs);
+	long long arc_num_max = (long long)(arc_max - arcs);
+	long long arc_num = (long long)(arc_last - arcs);
 	arc* arcs_old = arcs;
 
 	arc_num_max += arc_num_max / 2; if (arc_num_max & 1) arc_num_max ++;
